requirements.txt
spotipy>=2.22.1
requests>=2.31.0
urllib3>=2.0.0

Flask>=2.3.0
Werkzeug>=2.3.0

python-dotenv>=1.0.0
pydantic>=2.0.0

pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

orjson>=3.9.0

python-dateutil>=2.8.0

click>=8.1.0
rich>=13.0.0
colorama>=0.4.6

scikit-learn>=1.3.0
scipy>=1.11.0

pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.7.0
flake8>=6.0.0
mypy>=1.5.0

gunicorn>=21.2.0
waitress>=2.1.0

openpyxl>=3.1.0
xlsxwriter>=3.1.0
tabulate>=0.9.0

certifi>=2023.7.22
charset-normalizer>=3.2.0

# Remove discord.py unless you're building Discord integration
# discord.py>=2.3.0                # Discord bot integration (if needed)